requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
decorator==5.1.1
fastapi==0.75.0
h11==0.13.0
httptools==0.4.0
idna==3.3
jpath-finder==1.0.1
jsonpath-ng==1.5.3
ply==3.11
pydantic==1.9.0
python-dotenv==0.20.0
PyYAML==6.0
rstr==3.0.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.1
websockets==10.2